fix duration issue
