+++ app/routes/components/timer/Timer.js
@@ -3,6 +3,7 @@ import PropTypes from "prop-types";
 
 import { View, Text } from "react-native";
 
+import { msToTime } from "../../../store/services/helpers/moment";
 import styles from "./styles";
 
 class Timer extends Component {
@@ -10,20 +11,6 @@ class Timer extends Component {
     date: PropTypes.string.isRequired,
     timerStyle: View.propTypes.style.isRequired
   };
-
-  static msToTime(duration) {
-    let seconds = parseInt(duration / 1000 % 60, 10);
-    let minutes = parseInt(duration / (1000 * 60) % 60, 10);
-    let hours = parseInt(duration / (1000 * 60 * 60) % 24, 10);
-    let days = parseInt(duration / (1000 * 60 * 60 * 24) % 365, 10);
-
-    days = days < 10 ? `0${days}` : days;
-    hours = hours < 10 ? `0${hours}` : hours;
-    minutes = minutes < 10 ? `0${minutes}` : minutes;
-    seconds = seconds < 10 ? `0${seconds}` : seconds;
-
-    return `${days}:${hours}:${minutes}:${seconds}`;
-  }
 
   constructor(props) {
     super(props);
@@ -50,7 +37,7 @@ class Timer extends Component {
   render() {
     const { timerStyle } = this.props;
 
-    const time = Timer.msToTime(Math.abs(this.state.elapsedTime));
+    const time = msToTime(Math.abs(this.state.elapsedTime));
 
     return (
       <View style={[styles.container, timerStyle && timerStyle]}>
